requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.3

setuptools~=47.3.1
PyYAML~=5.3.1